Minor plotting changes

+++ plotting/make_plots.py
@@ -17,16 +17,25 @@ parser.add_argument("-e"   , "--era"   , type=int, default=2018  , help="era", r
 parser.add_argument('--doSyst', type=int, default=0, help="make systematic plots")
 parser.add_argument('--isMC', type=int, default=1, help="Is this MC or data")
 parser.add_argument('--blind', type=int, default=1, help="Blind the data (default=True)")
+parser.add_argument('--local', type=int, default=0, help="Local data or xrdcp from hadoop (default=False)")
 options = parser.parse_args()
+
+# other parameters for script
+username = getpass.getuser()
+if not options.local:
+    dataDir = "/mnt/T3_US_MIT/hadoop/scratch/{}/SUEP/{}/{}/".format(username,options.tag,options.dataset)
+else:
+    dataDir = "/work/submit/{}/SUEP/{}/{}/".format(username, options.tag, options.dataset)
+files = [file for file in os.listdir(dataDir)]
 
 # parameters for ABCD method
 var1_label = 'spher'
 var2_label = 'nconst'
 var1_val = 0.50
 var2_val = 25
-nbins = 100
-labels = ['ch']
-output_label = 'V6'
+nbins = 100                # applies to var1_label
+labels = ['ch']            # which selection to make plots for
+output_label = 'noPtCut'
 
 # cross section
 xsection = 1.0
@@ -38,11 +47,6 @@ with open('../data/xsections_{}.json'.format(options.era)) as file:
         xsection *= MC_xsecs[options.dataset]["br"]
     except:
         print("WARNING: I did not find the xsection for that MC sample. Check the dataset name and the relevant yaml file")
-
-#Get the list of files
-username = getpass.getuser()
-dataDir = "/mnt/T3_US_MIT/hadoop/scratch/{}/SUEP/{}/{}/".format(username,options.tag,options.dataset)
-files = [file for file in os.listdir(dataDir)]
 
 # output histos
 def create_output_file(l):
@@ -76,24 +80,21 @@ def create_output_file(l):
             "2D_rho1_nconst_"+label : Hist.new.Reg(100, 0, 20, name="rho1_"+label).Reg(99, 0, 200, name="nconst_"+label).Weight(),
             "2D_spher_ntracks_"+label : Hist.new.Reg(100, 0, 1.0, name="spher_"+label).Reg(200, 0, 500, name="ntracks_"+label).Weight(),
             "2D_spher_nconst_"+label : Hist.new.Reg(100, 0, 1.0, name="spher_"+label).Reg(99, 0, 200, name="nconst_"+label).Weight(),
-            "ht_" + label : Hist.new.Reg(1000, 0, 30000, name="ht_"+label, label='HT ' + label).Weight(),
+            "ht_" + label : Hist.new.Reg(1000, 0, 30000, name="ht_"+label, label='HT').Weight(),
+            "nJets_" + label : Hist.new.Reg(200, 0, 200, name="nJets_"+label, label='# Jets in Event').Weight(),
+            "nLostTracks_"+label : Hist.new.Reg(499, 0, 500, name=r + " nLostTracks_"+label, label=r + " # Lost Tracks in Event ").Weight(),
+            "2D_nJets_SUEPpT_"+label : Hist.new.Reg(200, 0, 200, name="nJets_"+label).Reg(99, 0, 200, name="nconst_"+label).Weight(),        
         
-            # region specific kinematic variables
-            "A_pt_"+label : Hist.new.Reg(100, 0, 2000, name="A pt_"+label, label=r"A $p_T$").Weight(),
-            "B_pt_"+label : Hist.new.Reg(100, 0, 2000, name="B pt_"+label, label=r"B $p_T$").Weight(),
-            "C_pt_"+label : Hist.new.Reg(100, 0, 2000, name="C pt_"+label, label=r"C $p_T$").Weight(),
-            "A_nconst_"+label : Hist.new.Reg(499, 0, 500, name="A nconst_"+label, label="A # Tracks in SUEP").Weight(),
-            "B_nconst_"+label : Hist.new.Reg(499, 0, 500, name="B nconst_"+label, label="B # Tracks in SUEP").Weight(),
-            "C_nconst_"+label : Hist.new.Reg(499, 0, 500, name="C nconst_"+label, label="C # Tracks in SUEP").Weight(),
-            "A_pt_nconst_"+label : Hist.new.Reg(100, 0, 2000, name="A pt_"+label).Reg(499, 0, 500, name="A nconst_"+label).Weight(),
-            "B_pt_nconst_"+label : Hist.new.Reg(100, 0, 2000, name="B pt_"+label).Reg(499, 0, 500, name="B nconst_"+label).Weight(),
-            "C_pt_nconst_"+label : Hist.new.Reg(100, 0, 2000, name="C pt_"+label).Reg(499, 0, 500, name="C nconst_"+label).Weight(),
-            "AB_pt_"+label : Hist.new.Reg(100, 0, 2000, name="AB pt_"+label, label=r"$p_T$").Weight(),
-            "AB_eta_"+label : Hist.new.Reg(100, -5, 5, name="AB eta_"+label, label=r"$\eta$").Weight(),
-            "AB_phi_"+label : Hist.new.Reg(100, 0, 6.5, name="AB phi_"+label, label=r"$\phi$").Weight(),
-            "AC_pt_"+label : Hist.new.Reg(100, 0, 2000, name="AC pt_"+label, label=r"$p_T$").Weight(),
-            "AC_eta_"+label : Hist.new.Reg(100, -5, 5, name="AC eta_"+label, label=r"$\eta$").Weight(),
-            "AC_phi_"+label : Hist.new.Reg(100, 0, 6.5, name="AC phi_"+label, label=r"$\phi$").Weight(),
+            # per region
+            for r in ["A", "B", "C"]:
+                r+"_pt_"+label : Hist.new.Reg(100, 0, 2000, name=r + "pt_"+label, label=r + r" $p_T$").Weight(),
+                r+"_nconst_"+label : Hist.new.Reg(499, 0, 500, name=r + " nconst_"+label, label=r + " # Tracks in SUEP").Weight(),
+                "2D_"+r+"_pt_nconst_"+label : Hist.new.Reg(100, 0, 2000, name=r+" pt_"+label).Reg(499, 0, 500, name=r+" nconst_"+label).Weight(),
+                r+"_eta_"+label : Hist.new.Reg(100, -5, 5, name=r+" eta_"+label, label=r + r"$\eta$").Weight()
+                r+"_phi_"+label : Hist.new.Reg(200, -6.5, 6.5, name=r + " phi_"+label, label=r + r"$\phi$").Weight(),
+                r +"_spher_"+label : Hist.new.Reg(100, 0, 1, name=r+"spher_"+label, label=r+"Sphericity").Weight(),
+                r + "_ntracks_"+label : Hist.new.Reg(499, 0, 500, name=r+"ntracks_"+label, label=r+"# Tracks in event").Weight(),
+        
     }
     if label == 'ch':# Christos only
         output2 = {
@@ -129,20 +130,25 @@ def h5load(ifile, label):
         return 0, 0
         
 # fill ABCD hists with dfs from hdf5 files
-frames = {"mult":[],"ch":[]}
 nfailed = 0
 weight = 0
 fpickle =  open("outputs/" + options.dataset+ "_" + output_label + '.pkl', "wb")
-output = {}
-for label in labels: output.update(create_output_file(label))
+output, sizeA, sizeC = {}
+for label in labels: 
+    output.update(create_output_file(label))
+    sizeA.update({label:0})
+    sizeC.update({label:0})
+    
 for ifile in tqdm(files):
     ifile = dataDir+"/"+ifile
 
-    if os.path.exists(options.dataset+'.hdf5'): os.system('rm ' + options.dataset+'.hdf5')
-    xrd_file = "root://t3serv017.mit.edu:/" + ifile.split('hadoop')[1]
-    os.system("xrdcp {} {}.hdf5".format(xrd_file, options.dataset))
-
-    df_vars, metadata = h5load(options.dataset+'.hdf5', 'vars')    
+    if not options.local:
+        if os.path.exists(options.dataset+'.hdf5'): os.system('rm ' + options.dataset+'.hdf5')
+        xrd_file = "root://t3serv017.mit.edu:/" + ifile.split('hadoop')[1]
+        os.system("xrdcp {} {}.hdf5".format(xrd_file, options.dataset))
+        df_vars, metadata = h5load(options.dataset+'.hdf5', 'vars')   
+    else:
+        df_vars, metadata = h5load(ifile, 'vars')   
 
     # check if file is corrupted, or empty
     if type(df_vars) == int: 
@@ -150,26 +156,30 @@ for ifile in tqdm(files):
         continue
     if df_vars.shape[0] == 0: continue
     
-    if options.isMC: weight += metadata['gensumweight']
+    # update the gensumweight
+    if options.isMC: weight[label] += metadata['gensumweight']
     
-    # store hts for the all event to be indexed
+    # store event-wide info to be indexed within each selection
     hts = df_vars['ht']
+    nJets = df_vars['ngood_fastjets']
+    nLostTracks = df_vars['nLostTracks']
     
     for label in labels:
-        df, metadata = h5load(options.dataset+'.hdf5', label) 
+        if not options.local: df, metadata = h5load(options.dataset+'.hdf5', label) 
+        else: df_vars, metadata = h5load(ifile, label)   
+        
         # parameters for ABCD plots
         var1 = 'SUEP_'+label+'_' + var1_label
         var2 = 'SUEP_'+label+'_' + var2_label
-
-        sizeA, sizeC = 0,0
                 
         # selections
         if var2_label == 'nconst': df = df.loc[df['SUEP_'+label+'_nconst'] >= 10]
         if var1_label == 'spher': df = df.loc[df['SUEP_'+label+'_spher'] >= 0.25]
         #df = df.loc[df['SUEP_'+label+'_pt'] >= 300]
+        
+        # blind
         if options.blind and not options.isMC:
              df = df.loc[((df[var1] < var1_val) & (df[var2] < var2_val)) | ((df[var1] >= var1_val) & (df[var2] < var2_val)) | ((df[var1] < var1_val) & (df[var2] >= var2_val))]
-        df = df.loc[df['SUEP_ch_pt'] >= 300]
 
         # divide the dfs by region
         df_A = df.loc[(df[var1] < var1_val) & (df[var2] < var2_val)]
@@ -177,8 +187,8 @@ for ifile in tqdm(files):
         df_C = df.loc[(df[var1] < var1_val) & (df[var2] >= var2_val)]
         df_D_obs = df.loc[(df[var1] >= var1_val) & (df[var2] >= var2_val)]
         
-        sizeC += df_C.shape[0]
-        sizeA += df_A.shape[0]
+        sizeC[label] += df_C.shape[0]
+        sizeA[label] += df_A.shape[0]
 
         # fill the ABCD histograms
         output["A_"+label].fill(df_A[var1])
@@ -186,8 +196,8 @@ for ifile in tqdm(files):
         output["D_exp_"+label].fill(df_B[var1])
         output["C_"+label].fill(df_C[var1])
         output["D_obs_"+label].fill(df_D_obs[var1])
-        output["ABCDvars_2D_"+label].fill(df[var1], df[var2])  
-
+        output["ABCDvars_2D_"+label].fill(df[var1], df[var2])
+        
         # fill the distributions as they are saved in the dataframes
         plot_labels = [key for key in df.keys() if key in list(output.keys())]
         for plot in plot_labels: output[plot].fill(df[plot])  
@@ -198,31 +208,31 @@ for ifile in tqdm(files):
         output["2D_rho1_nconst_"+label].fill(df["SUEP_"+label+"_rho1"], df["SUEP_"+label+"_nconst"])
         output["2D_spher_nconst_"+label].fill(df["SUEP_"+label+"_spher"], df["SUEP_"+label+"_nconst"])
         output["2D_spher_ntracks_"+label].fill(df["SUEP_"+label+"_spher"], df["SUEP_"+label+"_ntracks"])
-        output["AB_phi_"+label].fill(df['SUEP_' + label + '_phi'].loc[(df[var2] < var2_val)])
-        output["AB_eta_"+label].fill(df['SUEP_' + label + '_eta'].loc[(df[var2] < var2_val)])
-        output["AB_pt_"+label].fill(df['SUEP_' + label + '_pt'].loc[(df[var2] < var2_val)])
-        output["AC_phi_"+label].fill(df['SUEP_' + label + '_phi'].loc[(df[var1] < var1_val)])
-        output["AC_eta_"+label].fill(df['SUEP_' + label + '_eta'].loc[(df[var1] < var1_val)])
-        output["AC_pt_"+label].fill(df['SUEP_' + label + '_pt'].loc[(df[var1] < var1_val)])
-        output["A_pt_"+label].fill(df_A['SUEP_' + label + '_pt'])
-        output["B_pt_"+label].fill(df_B['SUEP_' + label + '_pt'])
-        output["C_pt_"+label].fill(df_C['SUEP_' + label + '_pt'])
-        output["A_nconst_"+label].fill(df_A['SUEP_' + label + '_nconst'])
-        output["B_nconst_"+label].fill(df_B['SUEP_' + label + '_nconst'])
-        output["C_nconst_"+label].fill(df_C['SUEP_' + label + '_nconst'])
-        output["A_pt_nconst_"+label].fill(df_A['SUEP_' + label + '_pt'], df_A['SUEP_' + label + '_nconst'])
-        output["B_pt_nconst_"+label].fill(df_B['SUEP_' + label + '_pt'], df_B['SUEP_' + label + '_nconst'])
-        output["C_pt_nconst_"+label].fill(df_C['SUEP_' + label + '_pt'], df_C['SUEP_' + label + '_nconst'])
         output["ht_" + label].fill(hts[df['SUEP_' + label + '_index']])
-    os.system('rm ' + options.dataset+'.hdf5')    
+        output["nJets_" + label].fill(nJets[df['SUEP_' + label + '_index']])
+        output["nLostTracks_" + label].fill(nLostTracks[df['SUEP_' + label + '_index']])
+        output["2D_nJets_SUEPpT_" + label].fill(nJets[df['SUEP_' + label + '_index']], df['SUEP_' + label + '_pt'])
         
-# ABCD method to obtain D expected
+        # per region
+        for r, df_r in zip(["A", "B", "C"], [df_A, df_B, df_C]):
+        
+            output[r + "_pt_"+label].fill(df_r['SUEP_' + label + '_pt'])
+            output[r + "_eta_"+label].fill(df_r['SUEP_' + label + '_eta'])
+            output[r + "_phi_"+label].fill(df_r['SUEP_' + label + '_phi'])
+            output[r + "_spher_"+label].fill(df_r['SUEP_' + label + '_spher'])
+            output[r + "_nconst_"+label].fill(df_r['SUEP_' + label + '_nconst'])
+            output[r + "_ntracks_"+label].fill(df_r['SUEP_' + label + '_ntracks'])
+            output["2D_" r + "_pt_nconst_"+label].fill(df_r['SUEP_' + label + '_pt'], df_r['SUEP_' + label + '_nconst'])
+    
+    if not options.local: os.system('rm ' + options.dataset+'.hdf5')    
+        
+# ABCD method to obtain D expected for each selection
 for label in labels:
     if sizeA>0.0:
-        CoverA =  sizeC / sizeA
+        CoverA =  sizeC[label] / sizeA[label]
     else:
         CoverA = 0.0
-        print("A region has no occupancy")
+        print("A region has no occupancy for selection", label)
     output["D_exp_"+label] = output["D_exp_"+label]*(CoverA)
     
 # apply normalization
@@ -235,6 +245,7 @@ else:
 pickle.dump(output, fpickle)
 print("Number of files that failed to be read:", nfailed)
 
+# save to root
 with uproot.recreate("outputs/" + options.dataset+ "_" + output_label + '.root') as froot:
-     for h, hist in output.items():
-         froot[h] = hist
+    for h, hist in output.items():
+        froot[h] = hist
